extract AssetInfoFileTestHelper
